feat(timescale): add every missing aggregate column in one upgrade pass
